♻️ refactor(frontend): seperate ui helper utils
