+++ frontend/src/util/ui.ts
@@ -1,0 +1,17 @@
+export const formatFileSize = (bytes: number): string => {
+  if (bytes === 0) return "0 Bytes";
+  const k = 1024;
+  const sizes = ["Bytes", "KB", "MB", "GB"];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + " " + sizes[i];
+};
+
+export const getFileIcon = (contentType: string): string => {
+  if (contentType.startsWith("image/")) return "🖼️";
+  if (contentType.startsWith("video/")) return "🎥";
+  if (contentType.startsWith("audio/")) return "🎵";
+  if (contentType.includes("pdf")) return "📕";
+  if (contentType.includes("text")) return "📄";
+  if (contentType.includes("zip") || contentType.includes("rar")) return "📦";
+  return "📄";
+};
